+++ src/queue/Queue.php
@@ -21,7 +21,6 @@ use yii\db\Expression;
 use yii\db\Query;
 use yii\di\Instance;
 use yii\mutex\Mutex;
-use yii\queue\cli\Signal;
 use yii\queue\ExecEvent;
 use yii\web\Response;
 
@@ -120,26 +119,37 @@ class Queue extends \yii\queue\cli\Queue implements QueueInterface
 
     /**
      * @inheritdoc
+     * @param bool $repeat Whether to continue listening when the queue is empty.
+     * @param int $timeout The number of seconds to wait between cycles
+     * @retrun int|null the exit code
      */
-    public function run()
+    public function run(bool $repeat = false, int $timeout = 0): ?int
     {
-        while (!Signal::isExit() && ($payload = $this->reserve())) {
-            if ($this->handleMessage($payload['id'], $payload['job'], $payload['ttr'], $payload['attempt'])) {
-                $this->release($payload['id']);
+        return $this->runWorker(function(callable $canContinue) use ($repeat, $timeout) {
+            while ($canContinue()) {
+                if ($payload = $this->reserve()) {
+                    if ($this->handleMessage($payload['id'], $payload['job'], $payload['ttr'], $payload['attempt'])) {
+                        $this->release($payload['id']);
+                    }
+                } else if (!$repeat) {
+                    break;
+                } else if ($timeout) {
+                    sleep($timeout);
+                }
             }
-        }
+        });
     }
 
     /**
      * Listens to the queue and runs new jobs.
      *
-     * @param integer $delay number of seconds for waiting new job.
+     * @param integer $timeout The number of seconds to wait between cycles
+     * @retrun int|null the exit code
+     * @deprecated in 3.6.11. Use [[run()]] instead.
      */
-    public function listen(int $delay)
+    public function listen(int $timeout = 0): ?int
     {
-        do {
-            $this->run();
-        } while (!$delay || sleep($delay) === 0);
+        return $this->run(true, $timeout);
     }
 
     /**
@@ -392,6 +402,7 @@ class Queue extends \yii\queue\cli\Queue implements QueueInterface
         $job = $this->serializer->unserialize($this->_jobData($result['job']));
 
         return ArrayHelper::filterEmptyStringsFromArray([
+            'delay' => max(0, $result['timePushed'] + $result['delay'] - time()),
             'status' => $this->_status($result),
             'error' => $result['error'] ?? '',
             'progress' => $result['progress'],
@@ -413,7 +424,6 @@ class Queue extends \yii\queue\cli\Queue implements QueueInterface
     {
         return $this->db->usePrimary(function() {
             return $this->_createJobQuery()
-                ->andWhere('[[timePushed]] <= :time - [[delay]]', [':time' => time()])
                 ->count('*', $this->db);
         });
     }
@@ -426,22 +436,23 @@ class Queue extends \yii\queue\cli\Queue implements QueueInterface
         // Move expired messages into waiting list
         $this->_moveExpired();
 
+        $query = $this->_createJobQuery();
+
         // Set up the reserved jobs condition
-        $reservedParams = [];
         $reservedCondition = $this->db->getQueryBuilder()->buildCondition([
             'and',
             ['fail' => false],
             ['not', ['timeUpdated' => null]],
-        ], $reservedParams);
+        ], $query->params);
 
-        $results = $this->db->usePrimary(function() use ($reservedCondition, $reservedParams, $limit) {
-            return $this->_createJobQuery()
-                ->select(['id', 'description', 'progress', 'progressLabel', 'timeUpdated', 'fail', 'error'])
-                ->andWhere('[[timePushed]] <= :time - [[delay]]', [':time' => time()])
-                ->orderBy(new Expression("CASE WHEN $reservedCondition THEN 1 ELSE 0 END DESC", $reservedParams))
-                ->addOrderBy(['priority' => SORT_ASC, 'id' => SORT_ASC])
-                ->limit($limit)
-                ->all($this->db);
+        $query
+            ->select(['id', 'description', 'timePushed', 'delay', 'progress', 'progressLabel', 'timeUpdated', 'fail', 'error'])
+            ->orderBy(new Expression("CASE WHEN $reservedCondition THEN 1 ELSE 0 END DESC"))
+            ->addOrderBy(['priority' => SORT_ASC, 'id' => SORT_ASC])
+            ->limit($limit);
+
+        $results = $this->db->usePrimary(function() use ($query) {
+            return $query->all($this->db);
         });
 
         $info = [];
@@ -449,6 +460,7 @@ class Queue extends \yii\queue\cli\Queue implements QueueInterface
         foreach ($results as $result) {
             $info[] = [
                 'id' => $result['id'],
+                'delay' => max(0, $result['timePushed'] + $result['delay'] - time()),
                 'status' => $this->_status($result),
                 'progress' => (int)$result['progress'],
                 'progressLabel' => $result['progressLabel'],
@@ -581,7 +593,8 @@ EOD;
             // Reserve one message
             $payload = $this->db->usePrimary(function() {
                 return $this->_createJobQuery()
-                    ->andWhere(['and', ['fail' => false, 'timeUpdated' => null], '[[timePushed]] <= :time - [[delay]]'], [':time' => time()])
+                    ->andWhere(['fail' => false, 'timeUpdated' => null])
+                    ->andWhere('[[timePushed]] + [[delay]] <= :time', ['time' => time()])
                     ->orderBy(['priority' => SORT_ASC, 'id' => SORT_ASC])
                     ->limit(1)
                     ->one($this->db);